youtube/bert_extractive_summarizer/requirements.txt
youtube-transcript-api==0.3.1
bert-extractive-summarizer==0.4.2
